refactor: move Volume and VolumeLevel to volume.rs
